+++ CHANGELOG.md
@@ -17,6 +17,19 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
 
 ### Added
 
+- Added `structlog.dev.get_active_console_renderer()` that returns the currently active `structlog.dev.ConsoleRenderer()`.
+  [#749](https://github.com/hynek/structlog/pull/749)
+
+- `structlog.dev.ConsoleRenderer()` now supports setting the `exception_formatter` attribute.
+
+  You can now disable the pretty-printing of exceptions by setting it to `structlog.dev.plain_traceback`:
+
+  ```python
+  cr = structlog.dev.get_active_console_renderer()
+  cr.exception_formatter = structlog.dev.plain_traceback
+  ```
+  [#749](https://github.com/hynek/structlog/pull/749)
+
 - Added `structlog.dev.ConsoleRenderer.get_default_column_styles` for reuse the default column styles.
   [#741](https://github.com/hynek/structlog/pull/741)